api/requirements.txt
Flask
requests
beautifulsoup4
gunicorn
flask_cors
python-dotenv
mysql-connector-python
Werkzeug
